requirements.txt
feedparser==6.0.11
openai>=1.0.0
python-dotenv==1.0.1
pyyaml==6.0.1
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.3
